Refactor road rendering and self-intersection checking
Move both away from using RectangleShapes and instead use vertices.
The visual of the road is now composed of a triangle strip, while the self-
intersection is checked using some points along the road, displaced to sides,
together with the DistanceField objects.

This improves both the rendering and road-generating performance.